Fixed minor import and version bug
