+++ pyproject.toml
@@ -20,9 +20,9 @@ robotpy_extras = [
 
 # Other pip packages to install
 requires = [
-     "robotpy-questnav==0.1.0",
+     "robotpy-questnav==0.1.1",
      "robotpy-commands-v2==2025.1.1",
-     "robotpy-rev==2025.0.2",
+     "robotpy-rev==2025.0.3.1",
      "numpy",
      "photonlibpy==2025.1.1",
      "robotpy-urcl==2025.0.1",
