+++ ultime/vision.py
@@ -13,8 +13,6 @@ from ultime.module import Module
 from ultime.timethis import tt
 
 
-
-
 april_tag_field_layout = AprilTagFieldLayout.loadField(
     AprilTagField.k2025ReefscapeWelded
 )
